requirements.txt
pandas>=2.0
python-dotenv>=1.0
tqdm>=4.66
openai>=1.40
gradio>=4.40
langchain-core>=0.3
langchain-openai>=0.2
langchain-chroma>=0.1.4
langchain-text-splitters>=0.3
gradio>=6.0
